pyproject.toml
[tool.poetry]
name = "notes-searcher"
version = "0.1.0"
description = ""
authors = ["Vicente Rocha <vicente.c.rocha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
lancedb = "^0.19.0"
sentence-transformers = "^3.4.1"
ollama = "^0.4.7"
python-frontmatter = "^1.1.0"
pandas = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
